Use `NOTION_HOME` for all sandboxing, reducing the amount of OS-specific logic. Fixes #207.
