+++ crates/notion-core/Cargo.toml
@@ -33,3 +33,4 @@ detect-indent = { git = "https://github.com/stefanpenner/detect-indent-rs", bran
 envoy = "0.1.3"
 mockito = { version = "0.14.0", optional = true }
 regex = "1.0.6"
+dirs = "1.0.4"
