+++ crates/notion-core/src/path/mod.rs
@@ -1,7 +1,10 @@
 //! Provides functions for determining the paths of files and directories
 //! in a standard Notion layout.
 
+use std::env;
 use std::path::{Path, PathBuf};
+
+use notion_fail::Fallible;
 
 cfg_if! {
     if #[cfg(feature = "universal-docs")] {
@@ -19,6 +22,90 @@ cfg_if! {
         mod windows;
         pub use self::windows::*;
     }
+}
+
+pub fn notion_home() -> Fallible<PathBuf> {
+    if let Some(home) = env::var_os("NOTION_HOME") {
+        Ok(Path::new(&home).to_path_buf())
+    } else {
+        default_notion_home()
+    }
+}
+
+pub fn cache_dir() -> Fallible<PathBuf> {
+    Ok(notion_home()?.join("cache"))
+}
+
+pub fn node_inventory_dir() -> Fallible<PathBuf> {
+    Ok(inventory_dir()?.join("node"))
+}
+
+pub fn yarn_inventory_dir() -> Fallible<PathBuf> {
+    Ok(inventory_dir()?.join("yarn"))
+}
+
+pub fn package_inventory_dir() -> Fallible<PathBuf> {
+    Ok(inventory_dir()?.join("packages"))
+}
+
+pub fn node_cache_dir() -> Fallible<PathBuf> {
+    Ok(cache_dir()?.join("node"))
+}
+
+pub fn node_index_file() -> Fallible<PathBuf> {
+    Ok(node_cache_dir()?.join("index.json"))
+}
+
+pub fn node_index_expiry_file() -> Fallible<PathBuf> {
+    Ok(node_cache_dir()?.join("index.json.expires"))
+}
+
+pub fn image_dir() -> Fallible<PathBuf> {
+    Ok(tools_dir()?.join("image"))
+}
+
+pub fn node_image_root_dir() -> Fallible<PathBuf> {
+    Ok(image_dir()?.join("node"))
+}
+
+pub fn node_image_dir(node: &str, npm: &str) -> Fallible<PathBuf> {
+    Ok(node_image_root_dir()?.join(node).join(npm))
+}
+
+pub fn yarn_image_root_dir() -> Fallible<PathBuf> {
+    Ok(image_dir()?.join("yarn"))
+}
+
+pub fn yarn_image_dir(version: &str) -> Fallible<PathBuf> {
+    Ok(yarn_image_root_dir()?.join(version))
+}
+
+pub fn yarn_image_bin_dir(version: &str) -> Fallible<PathBuf> {
+    Ok(yarn_image_dir(version)?.join("bin"))
+}
+
+pub fn shim_dir() -> Fallible<PathBuf> {
+    Ok(notion_home()?.join("bin"))
+}
+
+pub fn user_config_file() -> Fallible<PathBuf> {
+    Ok(notion_home()?.join("config.toml"))
+}
+
+pub fn tools_dir() -> Fallible<PathBuf> {
+    Ok(notion_home()?.join("tools"))
+}
+
+pub fn inventory_dir() -> Fallible<PathBuf> {
+    Ok(tools_dir()?.join("inventory"))
+}
+
+pub fn user_toolchain_dir() -> Fallible<PathBuf> {
+    Ok(tools_dir()?.join("user"))
+}
+
+pub fn user_platform_file() -> Fallible<PathBuf> {
+    Ok(user_toolchain_dir()?.join("platform.json"))
 }
 
 pub fn node_distro_file_name(version: &str) -> String {
