+++ crates/notion-core/src/path/unix.rs
@@ -1,11 +1,15 @@
 //! Provides functions for determining the paths of files and directories
 //! in a standard Notion layout in Unix-based operating systems.
 
-use std::{env, io};
+use std::io;
 use std::path::PathBuf;
 use std::os::unix;
 
+use dirs;
+
 use notion_fail::{ExitCode, Fallible, NotionFail};
+
+use super::{notion_home, node_image_dir, shim_dir};
 
 #[derive(Debug, Fail, NotionFail)]
 #[fail(display = "environment variable 'HOME' is not set")]
@@ -79,53 +83,13 @@ cfg_if! {
 //         launchscript                                    launchscript_file
 //         config.toml                                     user_config_file
 
-fn notion_home() -> Fallible<PathBuf> {
-    let home = env::home_dir().ok_or(NoHomeEnvVar)?;
+pub fn default_notion_home() -> Fallible<PathBuf> {
+    let home = dirs::home_dir().ok_or(NoHomeEnvVar)?;
     Ok(home.join(".notion"))
-}
-
-pub fn cache_dir() -> Fallible<PathBuf> {
-    Ok(notion_home()?.join("cache"))
-}
-
-pub fn node_inventory_dir() -> Fallible<PathBuf> {
-    Ok(inventory_dir()?.join("node"))
-}
-
-pub fn yarn_inventory_dir() -> Fallible<PathBuf> {
-    Ok(inventory_dir()?.join("yarn"))
-}
-
-pub fn package_inventory_dir() -> Fallible<PathBuf> {
-    Ok(inventory_dir()?.join("packages"))
-}
-
-pub fn node_cache_dir() -> Fallible<PathBuf> {
-    Ok(cache_dir()?.join("node"))
-}
-
-pub fn node_index_file() -> Fallible<PathBuf> {
-    Ok(node_cache_dir()?.join("index.json"))
-}
-
-pub fn node_index_expiry_file() -> Fallible<PathBuf> {
-    Ok(node_cache_dir()?.join("index.json.expires"))
 }
 
 pub fn archive_extension() -> String {
     String::from("tar.gz")
-}
-
-pub fn image_dir() -> Fallible<PathBuf> {
-    Ok(tools_dir()?.join("image"))
-}
-
-pub fn node_image_root_dir() -> Fallible<PathBuf> {
-    Ok(image_dir()?.join("node"))
-}
-
-pub fn node_image_dir(node: &str, npm: &str) -> Fallible<PathBuf> {
-    Ok(node_image_root_dir()?.join(node).join(npm))
 }
 
 pub fn node_image_bin_dir(node: &str, npm: &str) -> Fallible<PathBuf> {
@@ -137,28 +101,12 @@ pub fn node_image_3p_bin_dir(node: &str, npm: &str) -> Fallible<PathBuf> {
     Ok(node_image_dir(node, npm)?.join("lib/node_modules/.bin"))
 }
 
-pub fn yarn_image_root_dir() -> Fallible<PathBuf> {
-    Ok(image_dir()?.join("yarn"))
-}
-
-pub fn yarn_image_dir(version: &str) -> Fallible<PathBuf> {
-    Ok(yarn_image_root_dir()?.join(version))
-}
-
-pub fn yarn_image_bin_dir(version: &str) -> Fallible<PathBuf> {
-    Ok(yarn_image_dir(version)?.join("bin"))
+pub fn shim_file(toolname: &str) -> Fallible<PathBuf> {
+    Ok(shim_dir()?.join(toolname))
 }
 
 pub fn notion_file() -> Fallible<PathBuf> {
     Ok(notion_home()?.join("notion"))
-}
-
-pub fn shim_dir() -> Fallible<PathBuf> {
-    Ok(notion_home()?.join("bin"))
-}
-
-pub fn shim_file(toolname: &str) -> Fallible<PathBuf> {
-    Ok(shim_dir()?.join(toolname))
 }
 
 pub fn launchbin_file() -> Fallible<PathBuf> {
@@ -167,26 +115,6 @@ pub fn launchbin_file() -> Fallible<PathBuf> {
 
 pub fn launchscript_file() -> Fallible<PathBuf> {
     Ok(notion_home()?.join("launchscript"))
-}
-
-pub fn user_config_file() -> Fallible<PathBuf> {
-    Ok(notion_home()?.join("config.toml"))
-}
-
-pub fn tools_dir() -> Fallible<PathBuf> {
-    Ok(notion_home()?.join("tools"))
-}
-
-pub fn inventory_dir() -> Fallible<PathBuf> {
-    Ok(tools_dir()?.join("inventory"))
-}
-
-pub fn user_toolchain_dir() -> Fallible<PathBuf> {
-    Ok(tools_dir()?.join("user"))
-}
-
-pub fn user_platform_file() -> Fallible<PathBuf> {
-    Ok(user_toolchain_dir()?.join("platform.json"))
 }
 
 pub fn create_file_symlink(src: PathBuf, dst: PathBuf) -> Result<(), io::Error> {
