+++ tests/acceptance/support/sandbox.rs
@@ -359,13 +359,8 @@ impl SandboxBuilder {
 fn home_dir() -> PathBuf {
     paths::home()
 }
-#[cfg(unix)]
 fn notion_home() -> PathBuf {
     home_dir().join(".notion")
-}
-#[cfg(windows)]
-fn notion_home() -> PathBuf {
-    home_dir().join("AppData").join("Local").join("Notion")
 }
 fn notion_tmp_dir() -> PathBuf {
     notion_home().join("tmp")
@@ -439,9 +434,6 @@ impl Sandbox {
         let mut p = test_support::process::process(program);
         p.cwd(self.root())
             // sandbox the Notion environment
-            .env("NOTION_SANDBOX", "true") // used to indicate that Notion is running sandboxed, for directory logic in Windows
-            .env("HOME", home_dir())
-            .env("USERPROFILE", home_dir()) // windows
             .env("NOTION_HOME", notion_home())
             .env("PATH", &self.path)
             .env("NOTION_POSTSCRIPT", notion_postscript())
